requirements.txt
aiogram==3.13.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
aiosqlite==0.20.0
google-generativeai==0.8.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiohttp>=3.9.0,<3.11
